(refactor) switch helper function to a standalone js file
